fix: setuptips.py was not packaged to Pypi. We remove it and include the content in setup.py fix: add pandoc in .travis.yml to generate README clean: removed register.py
